Phase 7: Testing, bug fixes, and dead code removal
- Fix unfollow button visibility (showUnfollowButton/hideUnfollowButton calls)
- Fix context menu Follow action (was calling undefined toggleFollowMode)
- Remove ~230 lines of dead sprite sheet code (replaced by SVG system)
- Extract camera-rendering.js module (602 lines)
- Update README with final architecture table
- Update refactoring log with all bug fixes and dead code removal
- Add LOCAL_TESTING_SETUP.md for dev environment documentation

Bug fixes during Phase 7 testing:
- label undefined in addDistanceRings
- spriteMaterial undefined in createDistanceRingLabel
- filterSidebarTracks/updateMiniStatistics missing functions
- wasDragging → CameraState.wasDragging
- Camera angle formulas corrected in camera-rendering.js
- autoFadeTrails/trailFadeTime backward-compatible aliases
